+++ src/components/Tables/TopTokenTransactions/index.js
@@ -1,11 +1,11 @@
 import React, { useMemo } from 'react'
 import cx from 'classnames'
-import PanelWithHeader from '@santiment-network/ui/Panel/PanelWithHeader'
 import Panel from '@santiment-network/ui/Panel'
-import SmoothDropdown from '../../../components/SmoothDropdown/SmoothDropdown'
-import { COLUMNS, DEFAULT_SORTING } from './columns'
+import PanelWithHeader from '@santiment-network/ui/Panel/PanelWithHeader'
+import SmoothDropdown from '../../SmoothDropdown/SmoothDropdown'
 import Table from '../../../ducks/Table'
-import styles from './DetailedTransactionsTable.module.scss'
+import { COLUMNS, DEFAULT_SORTING } from './columns'
+import styles from './index.module.scss'
 
 const TransactionTable = ({
   header,
