+++ src/ducks/Studio/Widget/TopTransactionsTable/index.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect, useMemo } from 'react'
 import { useQuery } from '@apollo/react-hooks'
 import Icon from '@santiment-network/ui/Icon'
-import { TRANSACTIONS_QUERY } from './gql'
 import { newWidget } from '../utils'
 import Calendar from '../../AdvancedView/Calendar'
 import { DAY, getTimeIntervalFromToday } from '../../../../utils/dates'
 import HelpPopup from '../../../../components/HelpPopup/HelpPopup'
-import TransactionTable from '../../../../pages/Detailed/transactionsInfo/TransactionTable'
+import TransactionTable from '../../../../components/Tables/TopTokenTransactions'
+import { TRANSACTIONS_QUERY } from '../../../../components/Tables/TopTokenTransactions/gql'
 import { normalizeTransactionData } from '../../../../pages/Detailed/transactionsInfo/utils'
 import styles from './index.module.scss'
 import widgetStyles from '../Widget.module.scss'
