+++ src/ducks/UniswapProtocol/UniswapTopTransactions/UniswapTopTransactions.module.scss
@@ -29,8 +29,3 @@
   margin: 0 auto 0 16px;
   font-size: 5px;
 }
-
-.tableWrapper {
-  max-height: 800px;
-  overflow-y: auto;
-}
